+++ src/visualize/+mag/+graphics/+chart/+custom/Event.m
@@ -4,6 +4,8 @@ classdef Event < mag.graphics.chart.Chart
     properties
         % EVENTOFINTEREST Event name to plot.
         EventOfInterest (1, 1) string
+        % ENDTIME Final event end time.
+        EndTime datetime {mustBeScalarOrEmpty} = datetime.empty()
         % YOFFSET Offset of label describing y-axis value.
         YOffset (1, 1) double = 1
         % IGNOREMISSING Ignore missing values.
@@ -60,6 +62,12 @@ classdef Event < mag.graphics.chart.Chart
                 interestingEvents(locCombine, :) = [];
             end
 
+            if isempty(this.EndTime)
+                endTime = data.(data.Properties.DimensionNames{1})(end);
+            else
+                endTime = this.EndTime;
+            end
+
             time = interestingEvents.(interestingEvents.Properties.DimensionNames{1});
             variable = interestingEvents.(eventOfInterest);
 
@@ -69,7 +77,7 @@ classdef Event < mag.graphics.chart.Chart
 
             plotTime = repmat(datetime("now", TimeZone = "UTC"), 2 * numel(time), 1);
             plotTime(1:2:end) = time;
-            plotTime(2:2:end) = [time(2:end); data.(data.Properties.DimensionNames{1})(end)];
+            plotTime(2:2:end) = [time(2:end); endTime];
             plotTime = reshape(plotTime, 2, []);
 
             plotVariable = [variable, variable]';
@@ -82,7 +90,7 @@ classdef Event < mag.graphics.chart.Chart
             end
 
             % Plot vertical lines between mode changes.
-            xline(axes, [time; data.(data.Properties.DimensionNames{1})(end)], "--");
+            xline(axes, [time; endTime], "--");
 
             % Plot text annotation.
             if ~iscategorical(variable)